# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "title": "Phone Marketplace API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/api/phones": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "lists user's phones",
                "tags": [
                    "phones"
                ],
                "summary": "lists user's phones",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_is-web-y26_m3302-milovatskiy_internal_domain_general.Phone"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "adds a new phone listing",
                "tags": [
                    "phones"
                ],
                "summary": "adds a new phone listing",
                "parameters": [
                    {
                        "description": "Phone details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_is-web-y26_m3302-milovatskiy_internal_domain_general.Phone"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/api/phones/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "gets phone listing by id",
                "tags": [
                    "phones"
                ],
                "summary": "gets phone listing by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Phone ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Phone details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_is-web-y26_m3302-milovatskiy_internal_domain_general.Phone"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_is-web-y26_m3302-milovatskiy_internal_domain_general.Phone"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "modifies phone listing",
                "tags": [
                    "phones"
                ],
                "summary": "modifies phone listing",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Phone ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated phone details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_is-web-y26_m3302-milovatskiy_internal_domain_general.Phone"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_is-web-y26_m3302-milovatskiy_internal_domain_general.Phone"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "removes a phone listing",
                "tags": [
                    "phones"
                ],
                "summary": "removes a phone listing",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Phone ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_is-web-y26_m3302-milovatskiy_internal_domain_general.Phone": {
            "type": "object",
            "properties": {
                "Condition": {
                    "type": "string",
                    "enum": [
                        "Excellent",
                        "Good",
                        "Fair",
                        "Poor"
                    ],
                    "x-enum-varnames": [
                        "Excellent",
                        "Good",
                        "Fair",
                        "Poor"
                    ]
                },
                "battery": {
                    "type": "string"
                },
                "brand": {
                    "type": "string"
                },
                "camera": {
                    "type": "string"
                },
                "cpu": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isUsed": {
                    "type": "boolean"
                },
                "issues": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "screenSize": {
                    "type": "string"
                },
                "sellerID": {
                    "type": "integer"
                },
                "storage": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "\"Bearer\" followed by a space and the JWT token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}